% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html.R
\name{add_tag}
\alias{add_tag}
\title{Wrap string in html tag}
\usage{
add_tag(x, tag, args = "")
}
\arguments{
\item{x}{string}

\item{tag}{Tag to wrap the string in}

\item{args}{attritutes to add to the tag (i.e. class or css properties)}
}
\description{
Takes a string and adds a opening and closing html tag to it
i.e. add_tag('hi', 'h1') ->  <h1>hi</h1>
}
