% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffResult.R
\name{as.character.diffResult}
\alias{as.character.diffResult}
\title{as.character.diffResult}
\usage{
\method{as.character}{diffResult}(x, ...)
}
\arguments{
\item{x}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO
}
