% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{as_ascii_table}
\alias{as_ascii_table}
\title{as_ascii_table}
\usage{
as_ascii_table(dat, line_prefix = "  ")
}
\arguments{
\item{dat}{Input dataset to convert into a ascii table}

\item{line_prefix}{Symbols to prefix in front of every line of the table}
}
\description{
This function takes a \code{data.frame} and attempts to convert it into
a simple ascii format suitable for printing to the screen
It is assumed all variable values have a \code{as.character()} method
in order to cast them to character.
}
\keyword{internal}
