% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{as_cropped_char}
\alias{as_cropped_char}
\title{as_cropped_char}
\usage{
as_cropped_char(inval, crop_at = 30)
}
\arguments{
\item{inval}{a single element value}

\item{crop_at}{character limit}
}
\description{
Makes any character string above x chars
Reduce down to a x char string with ...
}
