% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{assert_valid_keys}
\alias{assert_valid_keys}
\title{Assert that keys are valid}
\usage{
assert_valid_keys(COMPARE, KEYS, component, msg)
}
\arguments{
\item{COMPARE}{(\code{list})\cr A named list of which each element is a \code{data.frame} with the
column \code{VARIABLE}}

\item{KEYS}{(\code{character})\cr name of key variables to check to make sure they don't contain
any issues}

\item{component}{(\code{character})\cr name of the component within \code{COMPARE} to check against}

\item{msg}{(\code{character})\cr error message to print if any of \code{KEYS} are found within
\code{COMPARE[component]$VARIABLE}}
}
\description{
Utility function to check that user provided "keys" aren't listed as a problem
variable of the current list of issues.
}
\keyword{internal}
