% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{attribute_breakdown}
\alias{attribute_breakdown}
\title{attribute_breakdown}
\usage{
attribute_breakdown(attkeep, attdrop, datin)
}
\arguments{
\item{attkeep}{att we're going to keep. Should be a char vector, either BASEatt or COMPatt}

\item{attdrop}{att we're going to drop. Should be a char vector, either BASEatt or COMPatt}

\item{datin}{The data frame being reshaped}
}
\description{
Split up an attribute data frame to just get base or compare
This is a convenience function to tidy up code
}
