% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{cast_variables}
\alias{cast_variables}
\title{cast_variables}
\usage{
cast_variables(BASE, COMPARE, SUPWARN = FALSE)
}
\arguments{
\item{BASE}{base dataset}

\item{COMPARE}{comparison dataset}

\item{SUPWARN}{Should warning about variables being cast be printed?}
}
\description{
Function to cast datasets columns if they have differing types
Restricted to specific cases, currently integer and double, and character and factor
}
