% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{cast_vector}
\alias{cast_vector}
\title{cast_vector}
\usage{
cast_vector(colin, typecast, colname, whichdat)
}
\arguments{
\item{colin}{column to cast}

\item{typecast}{type of casting to do}

\item{colname}{name of vector}

\item{whichdat}{whether base or compare is being casted (used for warnings)}
}
\description{
casts a vector depending on its type and input
}
