% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{class_adder}
\alias{class_adder}
\title{class_adder}
\usage{
class_adder(objectin, new_class, ...)
}
\arguments{
\item{objectin}{Object to add new class too}

\item{new_class}{New class to add to object}

\item{...}{Additional values (will be assigned as attributes)}
}
\description{
Adds a specific class to an object as well as adding any required
attribute data
}
