% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{compare_vectors.default}
\alias{compare_vectors.default}
\title{compare_vectors.default}
\usage{
\method{compare_vectors}{default}(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\description{
Default method, if the vector is not numeric or factor. Basic comparison
}
\keyword{internal}
