% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{compare_vectors.factor}
\alias{compare_vectors.factor}
\title{compare_vectors.factor}
\usage{
\method{compare_vectors}{factor}(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\description{
Compares factors. Sets them as character and then compares
}
