% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{compare_vectors.integer64}
\alias{compare_vectors.integer64}
\title{compare_vectors.int64}
\usage{
\method{compare_vectors}{integer64}(
  target,
  current,
  tolerance = sqrt(.Machine$double.eps),
  scale = NULL,
  ...
)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{tolerance}{Level of tolerance for differences between two variables}

\item{scale}{Scale that tolerance should be set on. If NULL assume absolute}

\item{...}{Not used}
}
\description{
Handle int64 vectors. Uses numeric comparison
}
\keyword{internal}
