% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_values.R
\name{compare_vectors.list}
\alias{compare_vectors.list}
\title{compare_vectors.list}
\usage{
\method{compare_vectors}{list}(target, current, ...)
}
\arguments{
\item{target}{base list}

\item{current}{a list to compare the target against}

\item{...}{Ignored}
}
\description{
This version compares the differences between lists relying on the identical function
}
