% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{construct_issue}
\alias{construct_issue}
\title{construct_issue}
\usage{
construct_issue(value, message, add_class = NULL)
}
\arguments{
\item{value}{the value of the object}

\item{message}{the value of the message attribute}

\item{add_class}{additional class to add}
}
\description{
Make an s3 object with class issue and possible additional class,
and assign other arguments to attributes
}
\keyword{internal}
