% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{construct_s3}
\alias{construct_s3}
\title{construct_s3}
\usage{
construct_s3(Class, ...)
}
\arguments{
\item{Class}{new class to provide to bject}

\item{....}{Other arguments}
}
\description{
Make an s3 object with added s3 class, preserving other arguments
}
