% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{convert_to_issue}
\alias{convert_to_issue}
\title{convert_to_issue}
\usage{
convert_to_issue(datin)
}
\arguments{
\item{datin}{data inputted}
}
\description{
converts the count value into the correct issue format
}
\keyword{internal}
