% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{crop_char_value}
\alias{crop_char_value}
\title{crop_char_value}
\usage{
crop_char_value(inval, crop_at = 30)
}
\arguments{
\item{inval}{a single element value}

\item{crop_at}{character limit}
}
\description{
Makes any character string above x chars
Reduce down to a x char string with ...
}
