% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{describe_dataframe}
\alias{describe_dataframe}
\title{Describe the datasets being compared}
\usage{
describe_dataframe(base, comp, base_name, comp_name)
}
\arguments{
\item{base}{(\code{data.frame})\cr base dataset to be described}

\item{comp}{(\code{data.frame})\cr comparison dataset to be described}

\item{base_name}{(\code{character})\cr name of the base dataset}

\item{comp_name}{(\code{character})\cr name of the comparison dataset}
}
\description{
This function is used to produce a basic summary table of the core
features of the two \code{data.frame}'s being compared.
}
\keyword{internal}
