% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDF.R
\name{dfdiff_has_issues}
\alias{dfdiff_has_issues}
\title{dfdiff_has_issues}
\usage{
dfdiff_has_issues(x)
}
\arguments{
\item{x}{dfdiff object}
}
\description{
Utility function which returns True if an dfdiff
object has  issues or False if an dfdiff object does not have issues
}
\examples{

# Example with no issues
x <- dfdiff( iris, iris )
dfdiff_has_issues(x)

# Example with issues
iris2 <- iris
iris2[2,2] <- NA
x <- dfdiff( iris , iris2 , suppress_warnings = TRUE)
dfdiff_has_issues(x)
}
