% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuerows.R
\name{dfdiff_issuerows}
\alias{dfdiff_issuerows}
\title{dfdiff_issuerows}
\usage{
dfdiff_issuerows(df, diff, vars = NULL)
}
\arguments{
\item{df}{dataframe to be subsetted}

\item{diff}{dfdiff object}

\item{vars}{(optional) character vector containing names of issue variables to subset dataframe 
on. A value of NULL (default) will be taken to mean available issue variables.}
}
\description{
This function takes a dfdiff object and a dataframe and subsets
the dataframe for problem rows as identified in the comparison object.
If \code{vars} has been specified only issue rows associated with those
variable(s) will be returned.
}
\details{
Note that dfdiff_issuerows can be used to subset against any dataframe. The only
requirement is that the original variables specified in the keys argument to dfdiff
are present on the dataframe you are subsetting against. However please note that if 
no keys are specified in dfdiff then rownumber is used which means using 
dfdiff_issuerows against an arbitary dataset can easily result in nonsense rows 
being returned. It is always recommended to supply keys to dfdiff.
}
\examples{
iris2 <- iris
for ( i in 1:3) iris2[i,i] <- 99
x <- dfdiff( iris , iris2, suppress_warnings = TRUE)
dfdiff_issuerows( iris , x)
dfdiff_issuerows( iris2 , x)
dfdiff_issuerows( iris2 , x , vars = "Sepal.Length")
dfdiff_issuerows( iris2 , x , vars = c("Sepal.Length" , "Sepal.Width"))
}
