% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{diffdf_get_options}
\alias{diffdf_get_options}
\title{diffdf_get_options}
\usage{
diffdf_get_options()
}
\value{
list of current diffdf options
}
\description{
Get the options for use in \code{\link{diffdf}}.
}
\examples{
 

diffdf_set_options(warnings = FALSE)
#' x <- subset(iris,  -Species)
x[1,2] <- 5
COMPARE <- diffdf(iris, x)

}
