% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depreciated.R
\name{diffdf_issuerows}
\alias{diffdf_issuerows}
\title{diffdf_issuerows}
\usage{
diffdf_issuerows(...)
}
\arguments{
\item{...}{Not Used}
}
\description{
This function has been depreciated and there is currently no replacement due to a perceived lack of usage. 
If you do wish to have a replacement created please create an enhancement ticket at: 
https://github.com/gowerc/diffdf/issues
}
