% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{diffdf_options_reset}
\alias{diffdf_options_reset}
\title{diffdf_options_reset}
\usage{
diffdf_options_reset()
}
\value{
invisibly returns default options
}
\description{
resets all diffdf options to default
}
\examples{
diffdf_options(warnings = FALSE)
diffdf_options_reset()
x <- subset(iris,  -Species)
x[1,2] <- 5
COMPARE <- diffdf(iris, x)

}
