% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{diffdf_set_options}
\alias{diffdf_set_options}
\title{diffdf_set_options}
\usage{
diffdf_set_options(warnings = TRUE, strict_numeric = TRUE,
  strict_factor = TRUE, file = NULL,
  tolerance = sqrt(.Machine$double.eps), scale = NULL)
}
\arguments{
\item{warnings}{\code{(logical)} Should diffdf display warnings when it finds issues?}

\item{strict_numeric}{Flag for strict numeric to numeric comparisons (default = TRUE). 
If False diffdf will cast integer to double where required for comparisons. 
Note that variables specified in the keys will never be casted.}

\item{strict_factor}{Flag for strict factor to character comparisons (default = TRUE). 
If False diffdf will cast factors to characters where required for comparisons. 
Note that variables specified in the keys will never be casted.}

\item{file}{Location and name of a text file to output the results to. 
Setting to NULL will cause no file to be produced.}

\item{tolerance}{Set tolerance for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance.}

\item{scale}{Set scale for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance.
Setting as NULL is a slightly more efficient version of scale = 1.}

\item{warnings}{Do you want to display warnings? (logical) (default = TRUE)}
}
\value{
diffdf options updated, and NULL returned
}
\description{
Set the options for use in \code{\link{diffdf}}. These are called in \code{diffdf(config = get_options())}.
Options which are not expressly given will be set to their default value by this function
}
\examples{
 

diffdf_set_options(warnings = FALSE)
#' x <- subset(iris,  -Species)
x[1,2] <- 5
COMPARE <- diffdf(iris, x)

}
