% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{diffopts}
\alias{diffopts}
\title{diffopts}
\usage{
diffopts(
  scale = 1,
  tolerance = sqrt(.Machine$double.eps),
  onfailure = "warning",
  ...
)
}
\arguments{
\item{scale}{Set scale for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance}

\item{tolerance}{Set tolerance for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance}

\item{onfailure}{How should messages be raised if a comparison fails. Options are "warning", "message", "error" & "nothing". Default = "warning"}

\item{...}{Not used but included for compatibility reasons}
}
\description{
Options to modify how diffdf performs its comparison
}
