% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{error_check_options}
\alias{error_check_options}
\title{error_check_options}
\usage{
error_check_options(funin, value)
}
\arguments{
\item{funin}{Function to check with}

\item{value}{Value to check meets expectations}
}
\description{
Executes check of value for given function
}
