% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{factor_to_character}
\alias{factor_to_character}
\title{factor_to_character}
\usage{
factor_to_character(dsin, vars = NULL)
}
\arguments{
\item{dsin}{input dataframe}

\item{vars}{variables to consider for conversion. Default NULL will consider
every variable within the dataset}
}
\description{
Takes a dataframe and converts any factor variables to character
}
\keyword{internal}
