% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{find_difference_wrapfun}
\alias{find_difference_wrapfun}
\title{find_difference_wrapfun}
\usage{
find_difference_wrapfun(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\description{
Wrapper which passes find_difference methods. Intended only for use
in is_different
}
