% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{first_class}
\alias{first_class}
\title{first_class}
\usage{
first_class(col)
}
\arguments{
\item{col}{Variable we want to get the class from}
}
\description{
Convenience function to grab the first class message from a class call
}
