% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_keyname.R
\name{generate_keyname}
\alias{generate_keyname}
\title{generate_keyname}
\usage{
generate_keyname(
  BASE,
  COMP,
  replace_names = c("..ROWNUMBER..", "..RN..", "..ROWN..", "..N..")
)
}
\arguments{
\item{BASE}{base dataset}

\item{COMP}{comparison dataset}

\item{replace_names}{a vector of replacement names. Used for recursion, should be edited in function for clarity}
}
\description{
Function to generate a name for the keys if not provided
}
