% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{get_casted_dataset}
\alias{get_casted_dataset}
\title{get_casted_dataset}
\usage{
get_casted_dataset(df, columns, whichdat)
}
\arguments{
\item{df}{dataset to be casted}

\item{columns}{columns to be casted}

\item{whichdat}{whether base or compare is being casted (used for messages)}
}
\description{
Internal utility function to loop across a dataset casting all target
variables
}
\keyword{internal}
