% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{get_casted_vector}
\alias{get_casted_vector}
\title{get_casted_vector}
\usage{
get_casted_vector(colin, colname, whichdat)
}
\arguments{
\item{colin}{column to cast}

\item{colname}{name of vector}

\item{whichdat}{whether base or compare is being casted (used for messages)}
}
\description{
casts a vector depending on its type and input
}
\keyword{internal}
