% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_column_mode}
\alias{get_column_mode}
\title{get_column_mode}
\usage{
get_column_mode(df, cols = NULL)
}
\arguments{
\item{df}{dataframe to get column modes from}

\item{cols}{columns to subset. Column modes will be returned in the same order as specified by cols.
If left as NULL it will return the model of all columns}
}
\description{
Convenience function to get the mode types of a dataframe.
}
