% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuerows.R
\name{get_issue_dataset}
\alias{get_issue_dataset}
\title{get_issue_dataset}
\usage{
get_issue_dataset(issue, diff)
}
\arguments{
\item{issue}{name of issue to extract issue from}

\item{diff}{dfdiff object which contains issues}
}
\description{
Internal function used by dfdiff_issuerows to extract the dataframe
from each a target issue. In particular it also strips off any 
non-key variables
}
