% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issue_message.default}
\alias{get_issue_message.default}
\title{get_issue_message.default}
\usage{
\method{get_issue_message}{default}(object)
}
\arguments{
\item{object}{inputted object}
}
\description{
Default method for geting issue message. Checks if the object has an issue
And returns the message. Returns a blank otherwise
}
