% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_issue_message}
\alias{get_issue_message}
\title{get_issue_message}
\usage{
get_issue_message(object, ...)
}
\arguments{
\item{object}{inputted object of class issue}

\item{...}{other arguments}
}
\description{
Simple function to grab the issue message
}
\keyword{internal}
