% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_print_message.issue}
\alias{get_print_message.issue}
\title{get_print_message.issue}
\usage{
\method{get_print_message}{issue}(object)
}
\arguments{
\item{object}{an object of class issue_basic}
}
\description{
Get text from a basic issue, based on the class of the value of the issue
}
