% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_properties}
\alias{get_properties}
\title{get_properties}
\usage{
get_properties(dsin)
}
\arguments{
\item{dsin}{input dataframe that you want to get the metadata from}
}
\description{
Returns a dataframe of metadata for a given dataset.
Returned values include variable names , class , mode , type & attributes
}
