% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{get_table}
\alias{get_table}
\title{get_table}
\usage{
get_table(dsin, row_limit = 10)
}
\arguments{
\item{dsin}{dataset}

\item{row_limit}{Maximum number of rows to display in difference tables.
Use \code{NULL} to show all rows. Default is 10.}
}
\description{
Generate nice looking table from a data frame
}
\keyword{internal}
