% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.default}
\alias{get_text.default}
\title{get_text.default}
\usage{
\method{get_text}{default}(object)
}
\arguments{
\item{object}{issue}
}
\description{
Errors, as this should only ever be given an issue
}
