% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.issue_basic}
\alias{get_text.issue_basic}
\title{get_text.issue_basic}
\usage{
\method{get_text}{issue_basic}(object)
}
\arguments{
\item{object}{an object of class issue_basic}
}
\description{
Get text from a basic issue, based on the class of the value of the issue
}
