% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.issue_count}
\alias{get_text.issue_count}
\title{get_text.issue_count}
\usage{
\method{get_text}{issue_count}(object)
}
\arguments{
\item{object}{an issue_vector object}
}
\description{
Generate text for an issue_count object, adds text at top then applies default method
}
