% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.issue_list}
\alias{get_text.issue_list}
\title{get_text.issue_list}
\usage{
\method{get_text}{issue_list}(object)
}
\arguments{
\item{object}{an object of class issue_list}
}
\description{
Get the text for a list object, by applying get_text over the list
}
