% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.issue_vector}
\alias{get_text.issue_vector}
\title{get_text.issue_vector}
\usage{
\method{get_text}{issue_vector}(object)
}
\arguments{
\item{object}{an issue_vector object}
}
\description{
Generate text for an issue_vector object, adds text at top then applies default method
}
