% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text}
\alias{get_text}
\title{get_text}
\usage{
get_text(object, ...)
}
\arguments{
\item{object}{inputted object of class issue}

\item{...}{other arguments}
}
\description{
Get the required text depending on type of issue
}
