% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{has_issue.default}
\alias{has_issue.default}
\title{has_issue.default}
\usage{
\method{has_issue}{default}(object)
}
\arguments{
\item{object}{Inputted object}
}
\description{
Default method for default objects, just returns true
}
