% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{has_issue.issue}
\alias{has_issue.issue}
\title{has_issue.issue}
\usage{
\method{has_issue}{issue}(object)
}
\arguments{
\item{object}{an object of class issue_basic}
}
\description{
Check whether a basic issue has some problems by checking nrows in the object
}
