% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{has_issue.issue_count}
\alias{has_issue.issue_count}
\title{has_issue.issue_count}
\usage{
\method{has_issue}{issue_count}(object)
}
\arguments{
\item{object}{an issue_count object}
}
\description{
Determine if a vector issue has any issues, by summing over value
}
