% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{has_issue}
\alias{has_issue}
\title{has_issue_message}
\usage{
has_issue(object, ...)
}
\arguments{
\item{object}{inputted object of class issue}

\item{...}{other arguments}
}
\description{
Determine if a function has an issue
}
