% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{has_unique_rows}
\alias{has_unique_rows}
\title{has_unique_rows}
\usage{
has_unique_rows(DAT, KEYS)
}
\arguments{
\item{DAT}{input data set (data frame)}

\item{KEYS}{Set of keys which should be unique}
}
\description{
Check if a data sets rows are unique
}
