% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_att_differences}
\alias{identify_att_differences}
\title{identify_att_differences}
\usage{
identify_att_differences(BASE, COMP, exclude_cols = "")
}
\arguments{
\item{BASE}{Base dataset for comparision (data.frame)}

\item{COMP}{Comparitor dataset to compare base against (data.frame)}

\item{exclude_cols}{Columns to exclude from comparision}
}
\description{
Identifys any attribute differences between two data frames
}
