% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_column_order_differences}
\alias{identify_column_order_differences}
\title{Find column ordering differences}
\usage{
identify_column_order_differences(BASE, COMP)
}
\arguments{
\item{BASE}{(\code{data.frame})\cr Base dataset for comparison}

\item{COMP}{(\code{data.frame})\cr Comparator dataset to compare base against}
}
\description{
Compares two datasets and outputs a table listing any differences in the column
orders between the two datasets. Columns that are not contained within both
are ignored however column ordering is derived prior to removing these columns.
}
\keyword{internal}
