% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_extra_cols}
\alias{identify_extra_cols}
\title{identify_extra_cols}
\usage{
identify_extra_cols(DS1, DS2)
}
\arguments{
\item{DS1}{Baseline dataset (data frame)}

\item{DS2}{Comparator dataset (data frame)}
}
\description{
Identifies columns that are in a baseline dataset but not in a comparator dataset
}
