% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_extra_rows}
\alias{identify_extra_rows}
\title{identify_extra_rows}
\usage{
identify_extra_rows(DS1, DS2, KEYS)
}
\arguments{
\item{DS1}{Baseline dataset (data frame)}

\item{DS2}{Comparator dataset (data frame)}

\item{KEYS}{List of variables that define a unique row within the datasets (strings)}
}
\description{
Identifies rows that are in a baseline dataset but not in a comparator dataset
}
