% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_matching_cols}
\alias{identify_matching_cols}
\title{identify_matching_cols}
\usage{
identify_matching_cols(DS1, DS2, EXCLUDE = "")
}
\arguments{
\item{DS1}{Input dataset 1 (data frame)}

\item{DS2}{Input dataset 2 (data frame)}

\item{EXCLUDE}{Columns to ignore}
}
\description{
Identifies columns with the same name in two data frames
}
