% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_mode_differences}
\alias{identify_mode_differences}
\title{identify_mode_differences}
\usage{
identify_mode_differences(BASE, COMP)
}
\arguments{
\item{BASE}{Base dataset for comparision (data.frame)}

\item{COMP}{Comparitor dataset to compare base against (data.frame)}
}
\description{
Identifys any mode differences between two data frames
}
