% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_properties}
\alias{identify_properties}
\title{identify_properties}
\usage{
identify_properties(dsin)
}
\arguments{
\item{dsin}{input dataframe that you want to get the metadata from}
}
\description{
Returns a dataframe of metadata for a given dataset.
Returned values include variable names , class , mode , type & attributes
}
