% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_unsupported_cols}
\alias{identify_unsupported_cols}
\title{identify_unsupported_cols}
\usage{
identify_unsupported_cols(dsin)
}
\arguments{
\item{dsin}{input dataset}
}
\description{
Identifies any columns for which the package is not setup to handle
}
