% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_variable_diff}
\alias{identify_variable_diff}
\title{identify_variable_diff}
\usage{
identify_variable_diff(VAR, DAT, KEYS, tolerance = sqrt(.Machine$double.eps),
  scale = NULL)
}
\arguments{
\item{VAR}{Variable to compare for differences (string)}

\item{DAT}{Input dataset (data_frame)}

\item{KEYS}{Variables which define a unique row within the dataset (strings)}

\item{tolerance}{Level of tolerance for numeric differences between two variables}

\item{scale}{Scale that tolerance should be set on. If NULL assume absolute}
}
\description{
For a given variable find all values between 2 datasets where they mismatch.
The function takes only a single dataset as an input which is expected to have the relevent
columns from 2 input datasets already merged together. It is expected that the variable names
are differeniated by .x and .y
}
\details{
Function is only expected to be called from identify_differences which highly formats
the input prior to passing it to this function
}
