% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{invert}
\alias{invert}
\title{invert}
\usage{
invert(x)
}
\arguments{
\item{x}{list}
}
\description{
Utility function used to replicated \code{purrr::transpose}. Turns a list inside
out.
}
\keyword{internal}
