% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_different.factor}
\alias{is_different.factor}
\title{is_different.factor}
\usage{
is_different.factor(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\description{
Compares factors. Sets them as character and then compares
}
