% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_different.numeric}
\alias{is_different.numeric}
\title{is_different.numeric}
\usage{
is_different.numeric(target, current, tolerance = sqrt(.Machine$double.eps),
  scale = NULL, ..., check.attributes = TRUE)
}
\arguments{
\item{target}{<<TODO>>}

\item{current}{<<TODO>>}

\item{tolerance}{<<TODO>>}

\item{scale}{<<TODO>>}

\item{...}{<<TODO>>}

\item{check.attributes}{<<TODO>>}
}
\description{
This is a modified version of the all.equal function
}
