% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_different_wrapfun}
\alias{is_different_wrapfun}
\title{is_different_wrapfun}
\usage{
is_different_wrapfun(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\description{
Wrapper which passes is_different methods. Intended only for use
in is_different
}
