% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issue}
\alias{issue}
\title{Issue Object}
\description{
Base class for containing issues
These are used to control the flagging and warning messages associated with 
issues found in the rcompare function
}
\section{Usage}{
 Not intended for direct use. This object is intended to be inhertied by other 
objects which call it in an apprioate manner.
}

\section{Arguments}{

\code{value} - The values to be used by \code{checkfun} when deciding if an issue has been found or not.
This is also what is printed out when the print method for RCOMPARE is called

\code{message} - Warning message to be displayed if an issue is found

\code{checkfun} - Function to be run on \code{value} which decides whether or not an issue has been 
found

\code{order} - The order in which the warning messages should be displayed in relation to the other
issues found
}

\section{Methods}{

\code{$get_print_message()} - Returns text for printing.
  
\code{$has_issue()}  - Returns T/F based upon whether \code{checkfun} finds an issue in \code{value}.

\code{$get_issue_message()} - Returns the \code{message} if there are any issues.
  
\code{$get_text(dsin , row_limit = 20)} - Primary formating function used to prepare \code{dsin} prior
to printing. This is intended to be called via the inherited objects rather than directly.
}

