% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issue_basic}
\alias{issue_basic}
\title{Issue Basic Object}
\description{
Basic issue - Inherits from Issue Object 
These are used to control the flagging and warning messages associated with 
issues found in the rcompare function
}
\section{Usage}{
  \code{ ob = issue_basic$new(...) }
}

\section{Methods}{

\code{$get_text()} - Gets formated text suitable for printing
\code{$has_issue()} - Returns T/F depending on whether the issue is valid
}

