% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{issue_vector}
\alias{issue_vector}
\title{Issue Vector Object}
\description{
Vector issue - Inherits from Issue Object 
These are used to control the flagging and warning messages associated with 
issues found in the rcompare function
}
\section{Usage}{
  \code{ ob = issue_vector$new(...) }
}

\section{Methods}{

\code{$get_text()} - Gets formated text suitable for printing
}

