% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_check.R
\name{make_check}
\alias{make_check}
\title{make a check for diffdf}
\usage{
make_check(check_fun, name, message)
}
\arguments{
\item{check_fun}{A function which conducts the check. This function should
take arguments base, comp, keys and opts. It must return a list with 
elements:
\itemize{
\item{result: }{A logical which indicates if the test is passed}
\item{data: }{The data returned by the test, to be passed to the user}
\item{display_title: }{ The title to be displayed when the output of 
 this test is printed}
\item{display_body: }{The body to be displayed when the output of this test 
  is printed. Empty list if test passed}
\item{exclude_cols: }{Optional vector of columns to exclude from future 
 tests}
\item{exclude_rows: }{Optional named list of rows to exclude in base 
 and compare. Each element should be a vector with row numbers to exclude}
}}

\item{name}{A name for this test. This will be used to identify it in the 
list returned to the user}

\item{message}{The message for diffdf to display if the test fails}
}
\description{
make a check for diffdf
}
\details{
Creates a check with the correct structure for diffdf. Function factory
 so argument inputs are contained in the function environment
}
