% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{make_pasteobject}
\alias{make_pasteobject}
\title{make_paste_object}
\usage{
make_pasteobject(dataframe_in, message, row_limit = 20)
}
\arguments{
\item{dataframe_in}{data frame to display}

\item{message}{Message which appears above data frame}

\item{row_limit}{This is the cut off point. Default = 20.}
}
\description{
Pastes together the message and the data frame.
If more than 20 rows, its truncated
If this is an attribute message, switches
the data frame into the truncated tibble version to aid reading
}
