% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{make_textout.rcompare_basic}
\alias{make_textout.rcompare_basic}
\title{make_textout.rcompare_basic}
\usage{
\method{make_textout}{rcompare_basic}(datin, ...)
}
\arguments{
\item{datin}{Input data frame}

\item{...}{any pass through options}
}
\description{
Function to convert object into ascii rendition.
Basic objects are passed along to make_textout.
}
