% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{make_textout.rcompare_list}
\alias{make_textout.rcompare_list}
\title{make_textout.rcompare_list}
\usage{
\method{make_textout}{rcompare_list}(datin)
}
\arguments{
\item{datin}{Input data frame}

\item{...}{any pass through options}
}
\description{
Function to convert object into ascii rendition.
<<TODO>>
}
