% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{make_textout}
\alias{make_textout}
\title{make_text out}
\usage{
make_textout(datin, ...)
}
\arguments{
\item{datin}{the data frame being tabulated}

\item{...}{additional arguments to pass through}
}
\description{
Attempts to convert an object into a ascii rendtion
suitable for printing or outputing to a file. This is the generic
function with methods being specified for specific object types
}
