% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{merge_options}
\alias{merge_options}
\title{merge_options}
\usage{
merge_options(opts = NULL, ...)
}
\arguments{
\item{opts}{List of named options}

\item{...}{named arguments}
}
\description{
Merges a named list of options with named arguments via dots
}
