% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{mod_stargazer}
\alias{mod_stargazer}
\title{mod_stargazer}
\usage{
mod_stargazer(...)
}
\arguments{
\item{...}{Any arguments to give to stargazer}
}
\description{
simple modification to return only the object, not print as well!
}
