% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dfdiff}
\alias{print.dfdiff}
\title{Print dfdiff objects}
\usage{
\method{print}{dfdiff}(x, ..., as_string = FALSE)
}
\arguments{
\item{x}{comparison object created by dfdiff().}

\item{...}{Additional arguments (not used)}

\item{as_string}{Return printed message as an R character vector?}
}
\description{
Print nicely formated version of an dfdiff object
}
\examples{
x <- subset( iris , -Species )
x[1,2] <- 5
COMPARE <- dfdiff( iris, x)
print( COMPARE )
print( COMPARE , "Sepal.Length" )
}
