% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.diffdf}
\alias{print.diffdf}
\title{Print diffdf objects}
\usage{
\method{print}{diffdf}(x, ..., as_string = FALSE, row_limit = 10)
}
\arguments{
\item{x}{comparison object created by diffdf().}

\item{...}{Additional arguments (not used)}

\item{as_string}{Return printed message as an R character vector?}

\item{row_limit}{Maximium number of rows to display per table (default = 10)}
}
\description{
Print nicely formatted version of an diffdf object
}
\examples{
x <- subset( iris , -Species )
x[1,2] <- 5
COMPARE <- diffdf( iris, x)
print( COMPARE )
print( COMPARE , "Sepal.Length" )
}
