% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffResult.R
\name{print.diffResult}
\alias{print.diffResult}
\title{print.diffResult}
\usage{
\method{print}{diffResult}(x, type = "ascii", rowlimit = 10, file = NULL, display = TRUE, ...)
}
\arguments{
\item{x}{TODO}

\item{type}{TODO}

\item{rowlimit}{TODO}

\item{file}{TODO}

\item{display}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO
}
