% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{produce_file}
\alias{produce_file}
\title{Produce_file}
\usage{
produce_file(outfile, COMPARE)
}
\arguments{
\item{outfile}{location to save file}

\item{COMPARE}{the compare object}
}
\description{
Outputs a text file based on the compare object
}
