% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{quickdf}
\alias{quickdf}
\title{quickdf}
\usage{
quickdf(l)
}
\arguments{
\item{l}{a list wwhich follows data frame requirements exactly}
}
\description{
Makes a data frame from a list very rapidly. Does not perform
Usual data.frame checks to allow for brevity
}
