% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare_has_issue}
\alias{rcompare_has_issue}
\title{rcompare_has_issue}
\usage{
rcompare_has_issue(x)
}
\arguments{
\item{x}{rcompare object}
}
\description{
Utility function which returns True if an rcompare
object has  issues or False if an rcompare object does not have issues
}
