% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare_has_issues}
\alias{rcompare_has_issues}
\title{rcompare_has_issues}
\usage{
rcompare_has_issues(x)
}
\arguments{
\item{x}{rcompare object}
}
\description{
Utility function which returns True if an rcompare
object has  issues or False if an rcompare object does not have issues
}
\examples{

# Example with no issues
x <- rcompare( iris, iris )
rcompare_has_issues(x)

# Example with issues
iris2 <- iris
iris2[2,2] <- NA
x <- rcompare( iris , iris2 , suppress_warnings = TRUE)
rcompare_has_issues(x)
}
