% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare_has_pass}
\alias{rcompare_has_pass}
\title{rcompare_has_pass}
\usage{
rcompare_has_pass(x)
}
\arguments{
\item{x}{rcompare object}
}
\description{
Utility function which returns True if an rcompare
object has no issues or False if an rcompare object does have issues
}
