% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_columns}
\alias{remove_columns}
\title{remove_columns}
\usage{
remove_columns(df, cols)
}
\arguments{
\item{df}{dataset to remove columns from}

\item{cols}{columns to remove (character string)}
}
\description{
Removes named columns from a dataset
}
