% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_rows}
\alias{remove_rows}
\title{remove_columns}
\usage{
remove_rows(df, rows)
}
\arguments{
\item{df}{dataset to remove rows from}

\item{rows}{rows to remove (index) (numeric vector)}
}
\description{
Removes rows from a dataset by index
}
