% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_ascii.R
\name{render_ascii}
\alias{render_ascii}
\title{file}
\description{
Method for writing ascii object to a file
}
\section{Super class}{
\code{\link[diffdf:renderAbstract]{diffdf::renderAbstract}} -> \code{render_ascii}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-file}{\code{render_ascii$file()}}
\item \href{#method-display}{\code{render_ascii$display()}}
\item \href{#method-h1}{\code{render_ascii$h1()}}
\item \href{#method-h2}{\code{render_ascii$h2()}}
\item \href{#method-h3}{\code{render_ascii$h3()}}
\item \href{#method-h4}{\code{render_ascii$h4()}}
\item \href{#method-p}{\code{render_ascii$p()}}
\item \href{#method-br}{\code{render_ascii$br()}}
\item \href{#method-table}{\code{render_ascii$table()}}
\item \href{#method-clone}{\code{render_ascii$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="diffdf" data-topic="renderAbstract" data-id="initialize">}\href{../../diffdf/html/renderAbstract.html#method-initialize}{\code{diffdf::renderAbstract$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="diffdf" data-topic="renderAbstract" data-id="not_yet_implemented">}\href{../../diffdf/html/renderAbstract.html#method-not_yet_implemented}{\code{diffdf::renderAbstract$not_yet_implemented()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-file"></a>}}
\if{latex}{\out{\hypertarget{method-file}{}}}
\subsection{Method \code{file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$file(filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{file to write to 
display

Method to print ascii object to the screen
h1

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-display"></a>}}
\if{latex}{\out{\hypertarget{method-display}{}}}
\subsection{Method \code{display()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$display()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h1"></a>}}
\if{latex}{\out{\hypertarget{method-h1}{}}}
\subsection{Method \code{h1()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$h1(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
h2

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h2"></a>}}
\if{latex}{\out{\hypertarget{method-h2}{}}}
\subsection{Method \code{h2()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$h2(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
h3

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h3"></a>}}
\if{latex}{\out{\hypertarget{method-h3}{}}}
\subsection{Method \code{h3()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$h3(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
h4

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-h4"></a>}}
\if{latex}{\out{\hypertarget{method-h4}{}}}
\subsection{Method \code{h4()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$h4(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
p

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-p"></a>}}
\if{latex}{\out{\hypertarget{method-p}{}}}
\subsection{Method \code{p()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$p(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
br

Convert string into ascii h1 format}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-br"></a>}}
\if{latex}{\out{\hypertarget{method-br}{}}}
\subsection{Method \code{br()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$br(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{string
table

This function takes a data.frame and attempts to convert it into
a simple ascii format suitable for printing to the screen
It is assumed all variable values have a as.character() method
in order to cast them to character.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table"></a>}}
\if{latex}{\out{\hypertarget{method-table}{}}}
\subsection{Method \code{table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$table(df, limitstring)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{Input dataset to convert into a ascii table}

\item{\code{limitstring}}{A string indicating whether the dataset has been subset}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{render_ascii$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
