% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{sort_df}
\alias{sort_df}
\title{sort_df}
\usage{
sort_df(df, by)
}
\arguments{
\item{df}{dataframe to be sorted}

\item{by}{variables to sort the dataframe by (character vector)}
}
\description{
Convenience function to sort a dataset (and hide the ugly syntax required to do so)
}
