% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{sort_then_join}
\alias{sort_then_join}
\title{sort_then_join}
\usage{
sort_then_join(string1, string2)
}
\arguments{
\item{string1}{first string}

\item{string2}{second string}
}
\description{
Convenience function to sort two strings and paste them together
}
