% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{string_pad}
\alias{string_pad}
\title{string_pad
 
Utility function used to replicate str_pad. Adds white space to either end 
of a string to get it to equal the desired length}
\usage{
string_pad(x, width)
}
\arguments{
\item{x}{string}

\item{width}{desired length}
}
\description{
string_pad
 
Utility function used to replicate str_pad. Adds white space to either end 
of a string to get it to equal the desired length
}
