% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{subset_se}
\alias{subset_se}
\title{subset_se}
\usage{
subset_se(df, rows = NULL, cols = NULL)
}
\arguments{
\item{df}{dataframe to subset}

\item{rows}{row names / index / lgl vector to keep}

\item{cols}{column names / lgl vector to keep}
}
\description{
A version of subset that works only on standard evaluation (to avoid un-expected issues)
}
