% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{valuefixer}
\alias{valuefixer}
\title{valuefixer}
\usage{
valuefixer(inval)
}
\arguments{
\item{inval}{a single element value}
}
\description{
Makes any character string above 20 chars
Reduce down to a 20 char string with ...
}
