% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_variables.R
\name{cast_variables}
\alias{cast_variables}
\title{cast_variables}
\usage{
cast_variables(BASE, COMPARE, keys)
}
\arguments{
\item{BASE}{base dataset}

\item{COMPARE}{comparison dataset}

\item{KEYS}{unique keys to compare with}
}
\description{
Function to cast datasets columns if they have differting types
Restricted to specific cases
}
