% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{compare_vectors.list}
\alias{compare_vectors.list}
\title{compare_vectors.list}
\usage{
\method{compare_vectors}{list}(target, current, ...)
}
\arguments{
\item{target}{the base list}

\item{current}{a list to compare base to}

\item{...}{Current unused}
}
\description{
This method of compare vectors aims to compare and find differences
between list objects.
}
