% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R, R/package.R
\docType{package}
\name{diffdf}
\alias{diffdf}
\title{diffdf}
\usage{
diffdf(base, comp, keys = NULL, opts = NULL, ...)
}
\arguments{
\item{base}{TODO}

\item{comp}{TODO}

\item{keys}{TODO}

\item{opts}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO

Functions for comparing two data.frames against 
each other. The core functionality is to provide a detailed breakdown of any differences 
between two data.frames as well as providing utility functions to help narrow down the 
source of problems and differences.
}
\author{
Craig Gower-Page
}
