% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R, R/package.R
\docType{package}
\name{diffdf}
\alias{diffdf}
\alias{diffdf}
\alias{diffdf-package}
\title{diffdf}
\usage{
diffdf(base, compare, keys = NULL, config = diffdf_options())
}
\arguments{
\item{base}{input dataframe}

\item{compare}{comparison dataframe}

\item{keys}{vector of variables (as strings) that defines a unique row in the base and compare dataframes}

\item{config}{A list containing options for diffdf. Will call \code{\link{diffdf_options}}.}
}
\description{
Compares 2 dataframes and outputs any differences.

Functions for comparing two data.frames against 
each other. The core functionality is to provide a detailed breakdown of any differences 
between two data.frames as well as providing utility functions to help narrow down the 
source of problems and differences.
}
\examples{
x <- subset( iris,  -Species)
x[1,2] <- 5
COMPARE <- diffdf( iris, x)
print( COMPARE )
print( COMPARE , "Sepal.Length" )

#### Sample data frames

DF1 <- data.frame(
    id = c(1,2,3,4,5,6),
    v1 = letters[1:6],
    v2 = c(NA , NA , 1 , 2 , 3 , NA)
)

DF2 <- data.frame(
    id = c(1,2,3,4,5,7),
    v1 = letters[1:6],
    v2 = c(NA , NA , 1 , 2 , NA , NA),
    v3 = c(NA , NA , 1 , 2 , NA , 4)
)

diffdf(DF1 , DF1 , keys = "id")

# We can control matching with scale/location for example:

DF1 <- data.frame(
    id = c(1,2,3,4,5,6),
    v1 = letters[1:6],
    v2 = c(1,2,3,4,5,6)
)
DF2 <- data.frame(
    id = c(1,2,3,4,5,6),
    v1 = letters[1:6],
    v2 = c(1.1,2,3,4,5,6)
)

diffdf(DF1 , DF2 , keys = "id")
diffdf_options(tolerance = 0.2)
diffdf(DF1 , DF2 , keys = "id")
diffdf_options(tolerance = 0.2, scale = 10)
diffdf(DF1 , DF2 , keys = "id")
 
# We can use strict_factor to compare factors with characters for example:

DF1 <- data.frame(
    id = c(1,2,3,4,5,6),
    v1 = letters[1:6],
    v2 = c(NA , NA , 1 , 2 , 3 , NA), 
    stringsAsFactors = FALSE
)

DF2 <- data.frame(
    id = c(1,2,3,4,5,6),
    v1 = letters[1:6],
    v2 = c(NA , NA , 1 , 2 , 3 , NA)
)
diffdf_options(strict_factor = TRUE)
diffdf(DF1 , DF2 , keys = "id")
diffdf_options(strict_factor = FALSE)
diffdf(DF1 , DF2 , keys = "id")
}
\author{
Craig Gower
}
