% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffdf.R
\name{diffdf_has_issues}
\alias{diffdf_has_issues}
\title{diffdf_has_issues}
\usage{
diffdf_has_issues(x)
}
\arguments{
\item{x}{diffdf object}
}
\description{
Utility function which returns TRUE if an diffdf
object has issues or FALSE if an diffdf object does not have issues
}
\examples{

# Example with no issues
x <- diffdf(iris, iris)
diffdf_has_issues(x)

# Example with issues
iris2 <- iris
iris2[2, 2] <- NA
x <- diffdf(iris, iris2, suppress_warnings = TRUE)
diffdf_has_issues(x)
}
