% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{diffdf_options}
\alias{diffdf_options}
\title{diffdf_options}
\usage{
diffdf_options(...)
}
\arguments{
\item{warnings}{\code{(logical)} Should diffdf display warnings when it finds issues? (default = TRUE)}

\item{strict_numeric}{Flag for strict numeric to numeric comparisons (default = TRUE). 
If False diffdf will cast integer to double where required for comparisons. 
Note that variables specified in the keys will never be casted.}

\item{strict_factor}{Flag for strict factor to character comparisons (default = TRUE). 
If False diffdf will cast factors to characters where required for comparisons. 
Note that variables specified in the keys will never be casted.}

\item{warnings}{Do you want to display warnings? (logical) (default = TRUE)}

\item{file}{Location and name of a text file to output the results to. 
Setting to NULL will cause no file to be produced. (Default = NULL)}

\item{tolerance}{Set tolerance for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance. (default = sqrt(.Machine$double.eps))}

\item{scale}{Set scale for numeric comparisons. Note that comparisons fail if (x-y)/scale > tolerance.
Setting as NULL is a slightly more efficient version of scale = 1. (default = NULL)}
}
\value{
invisibly returns options
}
\description{
Set options for diffdf. Will return the options, so can be called directly by diffdf.
}
\examples{
 

diffdf_options(warnings = FALSE)
#' x <- subset(iris,  -Species)
x[1,2] <- 5
COMPARE <- diffdf(iris, x)

}
