% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{diffopts}
\alias{diffopts}
\title{diffopts}
\usage{
diffopts(
  scale = 1,
  tolerance = sqrt(.Machine$double.eps),
  onfailure = "warning"
)
}
\arguments{
\item{scale}{TODO}

\item{tolerance}{TODO}

\item{onfailure}{TODO}
}
\description{
Description - TODO
}
