% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_print_message}
\alias{get_print_message}
\title{get_print_message}
\usage{
get_print_message(object, ...)
}
\arguments{
\item{object}{inputted object of class issue}

\item{...}{other arguments}
}
\description{
Get the required text depending on type of issue
}
