% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.default}
\alias{get_text.default}
\title{get_text.default}
\usage{
\method{get_text}{default}(object, dsin, row_limit = 10)
}
\arguments{
\item{object}{issue}

\item{dsin}{dataset provided by issue}

\item{row_limit}{Maximum number of rows displayed in dataset}
}
\description{
Default method for getting text, generates ascii output for an issue
}
