% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issues.R
\name{get_text.default}
\alias{get_text.default}
\title{Title}
\usage{
\method{get_text}{default}(object, dsin, row_limit = 10)
}
\arguments{
\item{object}{TODO}

\item{dsin}{TODO}

\item{row_limit}{TODO}
}
\description{
Description
}
