% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{identify_att_differences}
\alias{identify_att_differences}
\title{identify_att_differences}
\usage{
identify_att_differences(BASE, COMP, exclude_cols = "")
}
\arguments{
\item{BASE}{Base dataset for comparison (data.frame)}

\item{COMP}{Comparator dataset to compare base against (data.frame)}

\item{exclude_cols}{Columns to exclude from comparison}
}
\description{
Identifies any attribute differences between two data frames
}
