% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_different.default}
\alias{is_different.default}
\title{is_different.default}
\usage{
is_different.default(target, current, ...)
}
\arguments{
\item{target}{<<TODO>>}

\item{current}{<<TODO>>}

\item{...}{<<TODO>>}
}
\description{
<<TODO>>
}
