% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_different}
\alias{is_different}
\title{is_different}
\usage{
is_different(target, current, ...)
}
\arguments{
\item{target}{the base vector}

\item{current}{a vector to compare target to}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\value{
A boolean vector which is T if target and current are different
}
\description{
This determines if two vectors are different. It expects vectors of the same
length and type, and is intended to be used after checks have already been done
Initially picks out any nas (matching nas count as a match)
Then compares remaining vector
}
