% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_different.R
\name{is_variable_different}
\alias{is_variable_different}
\title{is_variable_different}
\usage{
is_variable_different(variablename, keynames, datain, ...)
}
\arguments{
\item{variablename}{name of variable being compared}

\item{keynames}{name of keys}

\item{datain}{Inputted dataset with base and compare vectors}

\item{...}{Additional arguments which might be passed through (numerical accuracy)}
}
\value{
A boolean vector which is T if target and current are different
}
\description{
This subsets the data set on the variable name, picks out differences and returns a tibble
of differences for the given variable
}
