% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{make_pasteobject}
\alias{make_pasteobject}
\title{make_paste_object}
\usage{
make_pasteobject(dataframe_in, message, att_expand, row_limit = 20)
}
\arguments{
\item{dataframe_in}{data frame to display}

\item{message}{Message which appears above data frame}

\item{att_expand}{T/F whether we should use the trunc_mat argument}

\item{row_limit}{This is the cut off point. Set at 20, but could be adjusted}
}
\description{
Pastes together the message and the data frame.
If more than 20 rows, its truncated
If this is an attribute message, switches
the data frame into the truncated tibble version to aid reading
}
