% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/producefile.R
\name{make_textout}
\alias{make_textout}
\title{make_text out}
\usage{
make_textout(fcompare, message, datin, att_expand = FALSE)
}
\arguments{
\item{fcompare}{The function used to check whether to display the item or not}

\item{message}{The message displayed above the column}

\item{datin}{the data frame being tabulated}

\item{att_expand}{false by default, whether we should have additional attribute breakdown}
}
\description{
Performs check as to whether the comparison is empty or not
If empty, returns null
if not, returns the data frame pasted with the message
If att_expand is true, will return breakdown of attributes as well
}
