% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.diffdf}
\alias{print.diffdf}
\title{Print diffdf objects}
\usage{
\method{print}{diffdf}(x, row_limit = 10, as_string = FALSE, ...)
}
\arguments{
\item{x}{comparison object created by diffdf().}

\item{row_limit}{Max row limit for difference tables (NULL to show all rows)}

\item{as_string}{Return printed message as an R character vector?}

\item{...}{Additional arguments (not used)}
}
\description{
Print nicely formatted version of an diffdf object
}
\examples{
x <- subset(iris, -Species)
x[1, 2] <- 5
COMPARE <- diffdf(iris, x)
print(COMPARE)
print(COMPARE, row_limit = 5)
}
