% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rcompare}
\alias{print.rcompare}
\title{Print nice formating of the rcompare object}
\usage{
\method{print}{rcompare}(x, VARIABLE = NULL, ...)
}
\arguments{
\item{x}{comparison object created by rcompare()}

\item{VARIABLE}{specific variable to inspect the differences of}

\item{...}{Additional arguments (not used)}
}
\description{
Print nice formating of the rcompare object
}
\examples{
# x <- iris[150,1]
# COMPARE <- rcompare( iris, x)
# print( COMPARE )
# print( COMPARE , "Sepal.Length" )
}
