% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rcompare}
\alias{print.rcompare}
\title{Print rcompare objects}
\usage{
\method{print}{rcompare}(x, VARIABLE = NULL, ...)
}
\arguments{
\item{x}{comparison object created by rcompare().}

\item{VARIABLE}{specific variable to inspect the differences of (string).}

\item{...}{Additional arguments (not used)}
}
\description{
Print nicely formated version of an rcompare object
}
\examples{
library(dplyr)
x <- iris \%>\% select( -Species)
x[1,2] <- 5
COMPARE <- rcompare( iris, x)
print( COMPARE )
print( COMPARE , "Sepal.Length" )
}
