% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare}
\alias{rcompare}
\title{Compares dataframes}
\usage{
rcompare(base, compare, keys = NULL, suppress_warnings = F)
}
\arguments{
\item{base}{input dataframe}

\item{compare}{comparison dataframe}

\item{keys}{vector of variables (as strings) that defines a unique row in the base and compare dataframes}

\item{suppress_warnings}{Do you want to suppress warnings? (logical)}
}
\description{
Emulates proc compare from SAS
Compares 2 data frames and outputs any differences.
}
\examples{
## compare( AAE , QC_AAE , c("USUBJID" , "AESEQ"))
}
