% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare}
\alias{rcompare}
\title{Compares dataframes}
\usage{
rcompare(BASE, COMP, KEY = NULL, SUPWARN = F)
}
\arguments{
\item{BASE}{input dataframe}

\item{COMP}{comparison dataframe}

\item{KEY}{list of variables (as strings) that defines a unique row in the dataframe}

\item{SUPWARN}{Do you want to suppress warnings}
}
\description{
Emulates proc compare from SAS
Compares 2 data frames and outputs any differences.
}
\examples{
## compare( AAE , QC_AAE , c("USUBJID" , "AESEQ"))
}
