% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompare.R
\name{rcompare}
\alias{rcompare}
\title{rcompare}
\usage{
rcompare(base, compare, keys = NULL, suppress_warnings = F,
  outfile = NULL)
}
\arguments{
\item{base}{input dataframe}

\item{compare}{comparison dataframe}

\item{keys}{vector of variables (as strings) that defines a unique row in the base and compare dataframes}

\item{suppress_warnings}{Do you want to suppress warnings? (logical)}

\item{outfile}{Location and name of a file to output the results to. Setting to NULL will cause no file to be produced.}
}
\description{
Compares 2 data frames and outputs any differences.
Emulates proc compare from SAS
}
\examples{
library(dplyr)
x <- iris \%>\% select( -Species)
x[1,2] <- 5
COMPARE <- rcompare( iris, x)
print( COMPARE )
print( COMPARE , "Sepal.Length" )

#### Example for ADaM VAD QC
# rcompare( AAE , QC_AAE , keys = c("USUBJID" , "AESEQ"))
}
