% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_tables.R
\name{recursive_reduce}
\alias{recursive_reduce}
\title{recursive_reduce
 
Utility function used to replicated purrr::reduce. Recursively applies a 
function to a list of elements until only 1 element remains}
\usage{
recursive_reduce(.l, .f)
}
\arguments{
\item{.l}{list of values to apply a function to}

\item{.f}{function to apply to each each element of the list in turn
i.e. .l[[1]] <- .f( .l[[1]] , .l[[2]]) ; .l[[1]] <- .f( .l[[1]] , .l[[3]])}
}
\description{
recursive_reduce
 
Utility function used to replicated purrr::reduce. Recursively applies a 
function to a list of elements until only 1 element remains
}
