% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffResult.R
\name{summary.diffResult}
\alias{summary.diffResult}
\title{summary.diffResult}
\usage{
\method{summary}{diffResult}(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{TODO}
}
\description{
Description - TODO
}
